feat(template-list): refactor background animation code
